pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "google-connections-mcp"
version = "0.1.0"
description = "MCP server for Google Workspace APIs (Sheets, Calendar, Gmail, Drive, Tasks)"
authors = [
    { name = "Michael Zrork", email = "michaelzrork@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "fastapi>=0.104.0",
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
    "uvicorn>=0.30.0",
    "starlette>=0.37.0",
    "pandas>=2.0.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
google-connections-mcp = "google_connections_mcp:main"

[project.urls]
"Homepage" = "https://github.com/michaelzrork/google-connections-mcp"
"Repository" = "https://github.com/michaelzrork/google-connections-mcp"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
